+++ src/Authentication/src/Credentials/Authenticators/IAuthenticator.php
@@ -28,5 +28,5 @@ interface IAuthenticator
      * @param string $error The error type, if there was one
      * @return bool True if the credential is authentic, otherwise false
      */
-    public function authenticate(ICredential $credential, ISubject &$subject = null, string &$error = null): bool;
+    public function tryAuthenticate(ICredential $credential, ISubject &$subject = null, string &$error = null): bool;
 }