+++ src/Authentication/tests/Credentials/Authenticators/JwtAuthenticatorTest.php
@@ -73,7 +73,7 @@ class JwtAuthenticatorTest extends TestCase
             ->willReturn(null);
         $subject = null;
         $error = null;
-        $this->assertFalse($this->authenticator->authenticate($credential, $subject, $error));
+        $this->assertFalse($this->authenticator->tryAuthenticate($credential, $subject, $error));
         $this->assertEquals(AuthenticatorErrorTypes::CREDENTIAL_MISSING, $error);
     }
 
@@ -85,7 +85,7 @@ class JwtAuthenticatorTest extends TestCase
             ->willReturn(false);
         $subject = null;
         $error = null;
-        $this->assertFalse($this->authenticator->authenticate($this->credential, $subject, $error));
+        $this->assertFalse($this->authenticator->tryAuthenticate($this->credential, $subject, $error));
         $this->assertEquals(AuthenticatorErrorTypes::CREDENTIAL_INCORRECT, $error);
     }
 
@@ -96,7 +96,7 @@ class JwtAuthenticatorTest extends TestCase
             ->method('verify')
             ->willReturn(true);
         $subject = null;
-        $this->assertTrue($this->authenticator->authenticate($this->credential, $subject));
+        $this->assertTrue($this->authenticator->tryAuthenticate($this->credential, $subject));
         /** @var ISubject $subject */
         $this->assertInstanceOf(ISubject::class, $subject);
         $this->assertEquals('Dave', $subject->getPrimaryPrincipal()->getId());