+++ src/Authentication/tests/Credentials/Authenticators/RefreshTokenAuthenticatorTest.php
@@ -81,7 +81,7 @@ class RefreshTokenAuthenticatorTest extends TestCase
             ->willReturn(null);
         $subject = null;
         $error = null;
-        $this->assertFalse($this->authenticator->authenticate($credential, $subject, $error));
+        $this->assertFalse($this->authenticator->tryAuthenticate($credential, $subject, $error));
         $this->assertEquals(AuthenticatorErrorTypes::CREDENTIAL_MISSING, $error);
     }
 
@@ -93,7 +93,7 @@ class RefreshTokenAuthenticatorTest extends TestCase
             ->willReturn(false);
         $subject = null;
         $error = null;
-        $this->assertFalse($this->authenticator->authenticate($this->credential, $subject, $error));
+        $this->assertFalse($this->authenticator->tryAuthenticate($this->credential, $subject, $error));
         $this->assertEquals(AuthenticatorErrorTypes::CREDENTIAL_INCORRECT, $error);
     }
 
@@ -108,7 +108,7 @@ class RefreshTokenAuthenticatorTest extends TestCase
             ->willReturn(false);
         $subject = null;
         $error = null;
-        $this->assertFalse($this->authenticator->authenticate($this->credential, $subject, $error));
+        $this->assertFalse($this->authenticator->tryAuthenticate($this->credential, $subject, $error));
         $this->assertEquals(AuthenticatorErrorTypes::CREDENTIAL_INCORRECT, $error);
     }
 
@@ -122,7 +122,7 @@ class RefreshTokenAuthenticatorTest extends TestCase
             ->method('has')
             ->willReturn(true);
         $subject = null;
-        $this->assertTrue($this->authenticator->authenticate($this->credential, $subject));
+        $this->assertTrue($this->authenticator->tryAuthenticate($this->credential, $subject));
         /** @var ISubject $subject */
         $this->assertInstanceOf(ISubject::class, $subject);
         $this->assertEquals('Dave', $subject->getPrimaryPrincipal()->getId());