+++ src/QueryBuilders/src/QueryBuilder.php
@@ -38,10 +38,10 @@ abstract class QueryBuilder
     /**
      * Starts a new select query
      *
-     * @param array $expression,... A variable list of select expressions
+     * @param string  ...$expression A variable list of select expressions
      * @return SelectQuery The select query builder
      */
-    abstract public function select(...$expression): SelectQuery;
+    abstract public function select(string ...$expression): SelectQuery;
 
     /**
      * Starts a new update query