+++ src/Views/src/Compilers/Php/PhpCompiler.php
@@ -30,7 +30,8 @@ class PhpCompiler implements ICompiler
     {
         $obStartLevel = ob_get_level();
         ob_start();
-        extract($view->getVars());
+        $vars = $view->getVars();
+        extract($vars);
 
         try {
             if (eval('?>' . $view->getContents()) === false) {
@@ -50,7 +51,7 @@ class PhpCompiler implements ICompiler
      *
      * @param Exception|Throwable $ex The exception to handle
      * @param int $obStartLevel The starting output buffer level
-     * @throws $ex Always rethrown
+     * @throws Throwable|Exception Always rethrown
      */
     protected function handleException($ex, int $obStartLevel): void
     {